part_4_langgraph/chapter_17_advanced_patterns/requirements.txt
# requirements.txt
langgraph==1.0.4
langchain-openai==1.1.1
python-dotenv==1.2.1
tenacity==9.1.2